+++ src/common/reviewCardForm.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import styles from "../common/styles.common/reviewCardForm.module.css";
 import { toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import addedToBascet from "../utils/addedToBascet";
+import addedToBascet from "../functions/addedToBascet";
 const ReviewCardForm = ({ name, price, _id, quantity, obj }) => {
   const [size, setSize] = useState(null);
   const handleClick = (e) => {
