+++ src/forms/login.jsx
@@ -1,0 +1,6 @@
+import LoginForm from "./loginForm";
+const Login = () => {
+  return <LoginForm />;
+};
+
+export default Login;
