+++ src/functions/filterSingleDataForSingleVieving.js
